Text file with gnu(n) for n<50000 from TUBS
Taken from http://www.icm.tu-bs.de/ag_algebra/software/small/number.html
by stripping off HTML code. Contains exactly one value per line. For the
value of gnu(n) see line number n.

As the description of the table at the above given URL says,
"The following table contains the numbers of isomorphism types
of groups of a given order for many orders up to 50.000. The
numbers of groups have been read from the Small Groups library
or calculated using the GrpConst package in GAP. There are 1082
orders for which the numbers of groups are not available in this
table (this is indicated by --). The smallest order for which
the number of groups is not known is 2048.